+++ spec/lib/waterdrop_spec.rb
@@ -1,3 +1,13 @@
 # frozen_string_literal: true
 
-# Nothing really here. Just a stub for Coditsu
+RSpec.describe_current do
+  subject(:waterdrop) { described_class }
+
+  describe '#gem_root' do
+    context 'when we want to get gem root path' do
+      let(:path) { Dir.pwd }
+
+      it { expect(waterdrop.gem_root.to_path).to eq path }
+    end
+  end
+end